小学期项目/finally! - 部署版/frontend/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
httpx>=0.25.0
pandas>=2.1.0
Pillow>=10.0.0
python-dotenv>=1.0.0
PyMySQL>=1.1.0
streamlit-lottie>=0.0.5
configparser>=6.0.0
